+++ yakssok/src/main/java/server/yakssok/domain/user/domain/repository/UserQueryRepository.java
@@ -1,4 +1,4 @@
-package server.yakssok.domain.user.repository;
+package server.yakssok.domain.user.domain.repository;
 
 import java.util.Optional;
